requirements.txt
blinker==1.7.0
certifi==2023.11.17
chardet==3.0.4
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
Flask==3.0.1
fonttools==4.47.2
fpdf==1.7.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
natural==0.2.0
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pluggy==1.4.0
pygame==2.5.2
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyparsing==3.1.1
PyPDF2==3.0.1
pytest==7.4.4
python-dateutil==2.8.2
pytz==2023.3.post1
rfc3986==1.5.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
text2num==2.5.0
threadpoolctl==3.2.0
tzdata==2023.4
Werkzeug==3.0.1
